Chat optimizations
